# Perf improvements for AvroDecimal / WriteResolver.ResolveDecimal