--- a/src/AvroConvert/AvroObjectServices/Schemas/AvroTypes/AvroDecimal.cs
+++ b/src/AvroConvert/AvroObjectServices/Schemas/AvroTypes/AvroDecimal.cs
@@ -55,19 +55,19 @@ namespace SolTechnology.Avro.AvroObjectServices.Schemas.AvroTypes
         /// <param name="value">The decimal value.</param>
         internal AvroDecimal(decimal value)
         {
-            var bytes = GetBytesFromDecimal(value);
-
-            var unscaledValueBytes = new byte[12];
-            Array.Copy(bytes, unscaledValueBytes, unscaledValueBytes.Length);
-
-            var unscaledValue = new BigInteger(unscaledValueBytes);
-            var scale = bytes[14];
-
-            if (bytes[15] == 128)
+            Span<byte> decimalByteBuffer = stackalloc byte[16];
+            WriteDecimalBytes(ref decimalByteBuffer, value);
+            
+#if NET6_0_OR_GREATER
+            var unscaledValue = new BigInteger(decimalByteBuffer.Slice(0, 12));
+#else
+            var unscaledValue = new BigInteger(decimalByteBuffer.Slice(0, 12).ToArray());
+#endif
+            if (decimalByteBuffer[15] == 128)
                 unscaledValue *= BigInteger.MinusOne;
 
             UnscaledValue = unscaledValue;
-            Scale = scale;
+            Scale = decimalByteBuffer[14];
             SeparatorCharacter = Separator;
         }
 
@@ -75,12 +75,25 @@ namespace SolTechnology.Avro.AvroObjectServices.Schemas.AvroTypes
         {
             SeparatorCharacter = Separator;
 
-            var unscaledValue = string.Join("", value.Split(SeparatorCharacter));
-            UnscaledValue = BigInteger.Parse(unscaledValue);
+            Span<char> buffer = stackalloc char[value.Length - 1];
+            GetValueWithoutSeparator(value, ref buffer, out var indexOfSeparatorCharacter);
 
-            var indexOfSeparatorCharacter = value.IndexOf(SeparatorCharacter);
+#if NET6_0_OR_GREATER
+            UnscaledValue = BigInteger.Parse(buffer);
+#else
+            UnscaledValue = BigInteger.Parse(buffer.ToString());
+#endif
             var scale = value.Length - indexOfSeparatorCharacter - 1;
             Scale = scale;
+
+            static void GetValueWithoutSeparator(string value, ref Span<char> buffer, out int indexOfSeparator)
+            {
+                indexOfSeparator = value.AsSpan().IndexOf(Separator);
+                
+                value.AsSpan(0, indexOfSeparator).CopyTo(buffer);
+                
+                value.AsSpan(indexOfSeparator + 1).CopyTo(buffer.Slice(2));
+            }
         }
 
         /// <summary>
@@ -778,6 +791,37 @@ namespace SolTechnology.Avro.AvroObjectServices.Schemas.AvroTypes
         public bool Equals(AvroDecimal other)
         {
             return Scale == other.Scale && UnscaledValue == other.UnscaledValue;
+        }
+        
+        private static void WriteDecimalBytes(ref Span<byte> buffer, decimal d)
+        {   
+#if NET6_0_OR_GREATER
+            Span<int> bits = stackalloc int[4];
+            decimal.GetBits(d, bits);
+#else
+            int[] bits = decimal.GetBits(d); 
+#endif  
+            int lo = bits[0];
+            int mid = bits[1];
+            int hi = bits[2];
+            int flags = bits[3];
+
+            buffer[0] = (byte)lo;
+            buffer[1] = (byte)(lo >> 8);
+            buffer[2] = (byte)(lo >> 0x10);
+            buffer[3] = (byte)(lo >> 0x18);
+            buffer[4] = (byte)mid;
+            buffer[5] = (byte)(mid >> 8);
+            buffer[6] = (byte)(mid >> 0x10);
+            buffer[7] = (byte)(mid >> 0x18);
+            buffer[8] = (byte)hi;
+            buffer[9] = (byte)(hi >> 8);
+            buffer[10] = (byte)(hi >> 0x10);
+            buffer[11] = (byte)(hi >> 0x18);
+            buffer[12] = (byte)flags;
+            buffer[13] = (byte)(flags >> 8);
+            buffer[14] = (byte)(flags >> 0x10);
+            buffer[15] = (byte)(flags >> 0x18);
         }
 
         private static byte[] GetBytesFromDecimal(decimal d)