This PR reduces allocations and adds more Span usage to `AvroDecimal` & `WriteResolver.ResolveDecimal`.

Using the following benchmark class

```
[MemoryDiagnoser]
public class WriterBenchmarksDecimal
{
    private decimal _value;
    private string _stringValue;

    private WriteResolver _resolver;
    private DecimalSchema _schema;
    private IWriter _writer;

    [GlobalSetup]
    public void Setup()
    {
        _value = 12345.6789m;
        _stringValue = "12345.6789";

        _resolver = new WriteResolver(null);
        _schema = (DecimalSchema)Schema.Create(_value);
        _writer = new Writer(Stream.Null);
    }

    [Benchmark]
    public int AvroDecimal_CreateFromDecimal() => new AvroDecimal(_value).Scale;

    [Benchmark]
    public int AvroDecimal_CreateFromString() => new AvroDecimal(_stringValue).Scale;

    [Benchmark]
    public int ResolveDecimal()
    {
        _resolver.ResolveDecimal(_schema, _value, _writer);
        return _schema.Scale;
    }
}
```

 running for optimized version vs current, these are the improvements

| Method                        | Version   | Mean      | Error    | StdDev   | Ratio  | Gen0   | Allocated | Allocation Ratio |
|-------------------------------|-----------|----------:|---------:|---------:|-------:|-------:|----------:|----------------:|
| AvroDecimal_CreateFromDecimal | Current   |  51.93 ns | 1.049 ns | 2.143 ns |   1.00 | 0.0363 |     152 B |            1.00 |
| AvroDecimal_CreateFromDecimal | Optimized |  20.20 ns | 0.112 ns | 0.099 ns |   0.39 |      - |         - |            0.00 |
| AvroDecimal_CreateFromString  | Current   | 214.11 ns | 4.088 ns | 6.364 ns |   1.00 | 0.0668 |     280 B |            1.00 |
| AvroDecimal_CreateFromString  | Optimized | 126.61 ns | 1.401 ns | 1.613 ns |   0.59 | 0.0248 |     104 B |            0.37 |
| ResolveDecimal                | Current   | 701.21 ns |14.053 ns |16.729 ns |   1.00 | 0.2060 |     864 B |            1.00 |
| ResolveDecimal                | Optimized | 464.97 ns | 5.368 ns | 4.759 ns |   0.66 | 0.1030 |     432 B |            0.50 |